requirements.txt
streamlit
pymupdf
google-generativeai
pandas
fpdf2
openpyxl
pillow
